Fix: Resolve unmapped property warning in ProcessoMapper
Added @Mapping(target = "participantes", ignore = true) to the toEntity method in ProcessoMapper to explicitly ignore the 'participantes' field during mapping from ProcessoDto to Processo. This resolves the compilation warning as the 'participantes' field is present in the Processo entity but not in the ProcessoDto, and is not intended to be mapped in this context. Also added the missing import for org.mapstruct.Mapping.